+++ api/index.php
@@ -10,6 +10,7 @@ require_once __DIR__ . '/core/helpers/date_helper.php';
 require_once __DIR__ . '/core/helpers/array_helper.php';
 require_once __DIR__ . '/core/helpers/string_helper.php';
 require_once __DIR__ . '/core/helpers/upload_helper.php';
+require_once __DIR__ . '/core/migration.php';
 
 Auth::init();